docker-airflow/retail_sales_etl/requirements.txt
apache-airflow==2.7.1
apache-airflow-providers-postgres==5.7.0
apache-airflow-providers-mysql==5.3.0
pandas==2.1.0
psycopg2-binary==2.9.7
mysql-connector-python==8.1.0
python-dotenv==1.0.0 